wip: refactor of args complete
