+++ crates/walrus-e2e-tests/tests/test_client.rs
@@ -142,7 +142,7 @@ where
         blobs_with_paths.push((path, data.to_vec()));
     }
 
-    let store_args = StoreArgs::default().with_store_optimizations(StoreOptimizations::none());
+    let store_args = StoreArgs::default().no_store_optimizations();
     let store_result = client
         .as_ref()
         .reserve_and_store_blobs_retry_committees_with_path(&blobs_with_paths, &store_args)
@@ -575,9 +575,8 @@ pub async fn test_store_and_read_duplicate_blobs() -> TestResult {
         blobs_with_paths.push((path, data.to_vec()));
     }
 
-    let store_args = StoreArgs::default().with_store_optimizations(StoreOptimizations::none());
+    let store_args = StoreArgs::default().no_store_optimizations();
     let store_result_with_path = client
-        .as_ref()
         .reserve_and_store_blobs_retry_committees_with_path(&blobs_with_paths, &store_args)
         .await?;
 
@@ -814,8 +813,8 @@ async fn test_delete_blob(blobs_to_create: u32) -> TestResult {
         let store_args = StoreArgs::default()
             .with_encoding_type(encoding_type)
             .with_epochs_ahead(idx)
-            .with_store_optimizations(StoreOptimizations::none())
-            .with_persistence(BlobPersistence::Deletable);
+            .no_store_optimizations()
+            .deletable();
         client
             .as_ref()
             .reserve_and_store_blobs(&blobs, &store_args)
@@ -825,7 +824,7 @@ async fn test_delete_blob(blobs_to_create: u32) -> TestResult {
     // Add a blob that is not deletable.
     let store_args = StoreArgs::default()
         .with_encoding_type(encoding_type)
-        .with_store_optimizations(StoreOptimizations::none());
+        .no_store_optimizations();
     let result = client
         .as_ref()
         .reserve_and_store_blobs(&blobs, &store_args)
@@ -868,10 +867,7 @@ async fn test_storage_nodes_do_not_serve_data_for_deleted_blobs() -> TestResult 
     let blob = walrus_test_utils::random_data(314);
     let blobs = vec![blob.as_slice()];
 
-    let store_args = StoreArgs::default()
-        .with_encoding_type(DEFAULT_ENCODING)
-        .with_store_optimizations(StoreOptimizations::none())
-        .with_persistence(BlobPersistence::Deletable);
+    let store_args = StoreArgs::default().no_store_optimizations().deletable();
     let results = client.reserve_and_store_blobs(&blobs, &store_args).await?;
     let store_result = results.first().expect("should have one blob store result");
     let blob_id = store_result
@@ -965,7 +961,7 @@ async fn test_store_quilt(blobs_to_create: u32) -> TestResult {
     let store_args = StoreArgs::default()
         .with_encoding_type(encoding_type)
         .with_epochs_ahead(2)
-        .with_store_optimizations(StoreOptimizations::none());
+        .no_store_optimizations();
     let store_operation_result = quilt_client
         .reserve_and_store_quilt::<QuiltVersionV1>(&quilt, &store_args)
         .await?;
@@ -1068,9 +1064,7 @@ async fn test_blocklist() -> TestResult {
     let client = client.as_ref();
     let blob = walrus_test_utils::random_data(314);
 
-    let store_args = StoreArgs::default()
-        .with_store_optimizations(StoreOptimizations::none())
-        .with_persistence(BlobPersistence::Deletable);
+    let store_args = StoreArgs::default().no_store_optimizations().deletable();
     let store_results = client
         .reserve_and_store_blobs(&[&blob], &store_args)
         .await?;
@@ -1156,7 +1150,7 @@ async fn test_blob_operations_with_credits() -> TestResult {
     // Store a blob with credits
     let blob_data = walrus_test_utils::random_data(314);
     let blobs = vec![blob_data.as_slice()];
-    let store_args = StoreArgs::default().with_store_optimizations(StoreOptimizations::none());
+    let store_args = StoreArgs::default().no_store_optimizations();
     let store_result = client.reserve_and_store_blobs(&blobs, &store_args).await?;
 
     let blob_object = match &store_result[0] {
@@ -1431,7 +1425,7 @@ async fn test_burn_blobs() -> TestResult {
     let mut blob_object_ids = vec![];
     for idx in 0..N_BLOBS {
         let blob = walrus_test_utils::random_data(314 + idx);
-        let store_args = StoreArgs::default().with_store_optimizations(StoreOptimizations::none());
+        let store_args = StoreArgs::default().no_store_optimizations();
         let result = client
             .as_ref()
             .reserve_and_store_blobs(&[blob.as_slice()], &store_args)
@@ -1480,9 +1474,8 @@ async fn test_extend_owned_blobs() -> TestResult {
 
     let current_epoch = client.as_ref().sui_client().current_epoch().await?;
     let blob = walrus_test_utils::random_data(314);
-    let encoding_type = DEFAULT_ENCODING;
 
-    let store_args = StoreArgs::default().with_store_optimizations(StoreOptimizations::none());
+    let store_args = StoreArgs::default().no_store_optimizations();
 
     let result = client
         .as_ref()
@@ -1544,7 +1537,7 @@ async fn test_share_blobs() -> TestResult {
         test_cluster::E2eTestSetupBuilder::new().build().await?;
 
     let blob = walrus_test_utils::random_data(314);
-    let store_args = StoreArgs::default().with_store_optimizations(StoreOptimizations::none());
+    let store_args = StoreArgs::default().no_store_optimizations();
     let result = client
         .as_ref()
         .reserve_and_store_blobs(&[blob.as_slice()], &store_args)
@@ -1644,7 +1637,7 @@ async fn test_post_store_action(
     let blob_data = walrus_test_utils::random_data_list(314, 4);
     let blobs: Vec<&[u8]> = blob_data.iter().map(AsRef::as_ref).collect();
     let store_args = StoreArgs::default()
-        .with_store_optimizations(StoreOptimizations::none())
+        .no_store_optimizations()
         .with_post_store(post_store);
     let results = client
         .as_ref()
@@ -1811,7 +1804,7 @@ async fn test_quorum_contract_upgrade() -> TestResult {
     // Store a blob after the upgrade to check if everything works after the upgrade.
     let blob_data = walrus_test_utils::random_data_list(314, 1);
     let blobs: Vec<&[u8]> = blob_data.iter().map(AsRef::as_ref).collect();
-    let store_args = StoreArgs::default().with_store_optimizations(StoreOptimizations::none());
+    let store_args = StoreArgs::default().no_store_optimizations();
     let _results = client
         .as_ref()
         .reserve_and_store_blobs_retry_committees(&blobs, &store_args)
@@ -1948,7 +1941,7 @@ impl<'a> BlobAttributeTestContext<'a> {
         for idx in 1..blobs_to_create + 1 {
             let store_args = StoreArgs::default()
                 .with_epochs_ahead(idx)
-                .with_store_optimizations(StoreOptimizations::none());
+                .no_store_optimizations();
             client
                 .as_mut()
                 .reserve_and_store_blobs(&blobs, &store_args)
